@@ -16,6 +16,7 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
+import asyncio
 import ipaddress
 import logging
 import socket
@@ -34,8 +35,17 @@ except ImportError as e:
 log = logging.getLogger(__name__)
 
 
-class TCP(socks.socksocket):
+class TCP:
+    TIMEOUT = 10
+
     def __init__(self, ipv6: bool, proxy: dict):
+        self.socket = None
+
+        self.reader = None  # type: asyncio.StreamReader
+        self.writer = None  # type: asyncio.StreamWriter
+
+        self.lock = asyncio.Lock()
+
         if proxy.get("enabled", False):
             hostname = proxy.get("hostname", None)
             port = proxy.get("port", None)
@@ -43,14 +53,14 @@ class TCP(socks.socksocket):
             try:
                 ip_address = ipaddress.ip_address(hostname)
             except ValueError:
-                super().__init__(socket.AF_INET)
+                self.socket = socks.socksocket(socket.AF_INET)
             else:
                 if isinstance(ip_address, ipaddress.IPv6Address):
-                    super().__init__(socket.AF_INET6)
+                    self.socket = socks.socksocket(socket.AF_INET6)
                 else:
-                    super().__init__(socket.AF_INET)
+                    self.socket = socks.socksocket(socket.AF_INET)
 
-            self.set_proxy(
+            self.socket.set_proxy(
                 proxy_type=socks.SOCKS5,
                 addr=hostname,
                 port=port,
@@ -58,37 +68,52 @@ class TCP(socks.socksocket):
                 password=proxy.get("password", None)
             )
 
-            log.info("Using proxy {}:{}".format(hostname, port))
+            log.info(f"Using proxy {hostname}:{port}")
         else:
-            super().__init__(
+            self.socket = socks.socksocket(
                 socket.AF_INET6 if ipv6
                 else socket.AF_INET
             )
 
-        self.settimeout(10)
+        self.socket.settimeout(TCP.TIMEOUT)
+
+    async def connect(self, address: tuple):
+        self.socket.connect(address)
+        self.reader, self.writer = await asyncio.open_connection(sock=self.socket)
 
     def close(self):
         try:
-            self.shutdown(socket.SHUT_RDWR)
-        except OSError:
-            pass
-        finally:
-            # A tiny sleep placed here helps avoiding .recv(n) hanging until the timeout.
-            # This is a workaround that seems to fix the occasional delayed stop of a client.
-            time.sleep(0.001)
-            super().close()
+            self.writer.close()
+        except AttributeError:
+            try:
+                self.socket.shutdown(socket.SHUT_RDWR)
+            except OSError:
+                pass
+            finally:
+                # A tiny sleep placed here helps avoiding .recv(n) hanging until the timeout.
+                # This is a workaround that seems to fix the occasional delayed stop of a client.
+                time.sleep(0.001)
+                self.socket.close()
 
-    def recvall(self, length: int) -> bytes or None:
+    async def send(self, data: bytes):
+        async with self.lock:
+            self.writer.write(data)
+            await self.writer.drain()
+
+    async def recv(self, length: int = 0):
         data = b""
 
         while len(data) < length:
             try:
-                packet = super().recv(length - len(data))
-            except OSError:
+                chunk = await asyncio.wait_for(
+                    self.reader.read(length - len(data)),
+                    TCP.TIMEOUT
+                )
+            except (OSError, asyncio.TimeoutError):
                 return None
             else:
-                if packet:
-                    data += packet
+                if chunk:
+                    data += chunk
                 else:
                     return None
 
