@@ -16,24 +16,22 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with Pyrogram.  If not, see <http://www.gnu.org/licenses/>.
 
+import asyncio
 import logging
-import threading
-import time
-from datetime import timedelta, datetime
-from hashlib import sha1, sha256
+import os
+from concurrent.futures.thread import ThreadPoolExecutor
+from datetime import datetime, timedelta
+from hashlib import sha1
 from io import BytesIO
-from os import urandom
-from queue import Queue
-from threading import Event, Thread
 
 import pyrogram
 from pyrogram import __copyright__, __license__, __version__
-from pyrogram.api import functions, types, core
-from pyrogram.api.all import layer
-from pyrogram.api.core import Message, TLObject, MsgContainer, Long, FutureSalt, Int
+from pyrogram import raw
 from pyrogram.connection import Connection
-from pyrogram.crypto import AES, KDF
+from pyrogram.crypto import mtproto
 from pyrogram.errors import RPCError, InternalServerError, AuthKeyDuplicated, FloodWait
+from pyrogram.raw.all import layer
+from pyrogram.raw.core import TLObject, MsgContainer, Int, FutureSalt, FutureSalts
 from .internals import MsgId, MsgFactory
 
 log = logging.getLogger(__name__)
@@ -42,18 +40,18 @@ log = logging.getLogger(__name__)
 class Result:
     def __init__(self):
         self.value = None
-        self.event = Event()
+        self.event = asyncio.Event()
 
 
 class Session:
     INITIAL_SALT = 0x616e67656c696361
-    NET_WORKERS = 1
     START_TIMEOUT = 1
     WAIT_TIMEOUT = 15
-    SLEEP_THRESHOLD = 60
+    SLEEP_THRESHOLD = 10
     MAX_RETRIES = 5
     ACKS_THRESHOLD = 8
     PING_INTERVAL = 5
+    EXECUTOR_SIZE_THRESHOLD = 512
 
     notice_displayed = False
 
@@ -71,22 +69,26 @@ class Session:
         64: "[64] invalid container"
     }
 
+    executor = ThreadPoolExecutor(2, thread_name_prefix="CryptoWorker")
+
     def __init__(
         self,
-        client: pyrogram,
+        client: "pyrogram.Client",
         dc_id: int,
         auth_key: bytes,
+        test_mode: bool,
         is_media: bool = False,
         is_cdn: bool = False
     ):
         if not Session.notice_displayed:
-            print("Pyrogram v{}, {}".format(__version__, __copyright__))
-            print("Licensed under the terms of the " + __license__, end="\n\n")
+            print(f"Pyrogram v{__version__}, {__copyright__}")
+            print(f"Licensed under the terms of the {__license__}", end="\n\n")
             Session.notice_displayed = True
 
         self.client = client
         self.dc_id = dc_id
         self.auth_key = auth_key
+        self.test_mode = test_mode
         self.is_media = is_media
         self.is_cdn = is_cdn
 
@@ -94,68 +96,60 @@ class Session:
 
         self.auth_key_id = sha1(auth_key).digest()[-8:]
 
-        self.session_id = Long(MsgId())
+        self.session_id = os.urandom(8)
         self.msg_factory = MsgFactory()
 
         self.current_salt = None
 
         self.pending_acks = set()
 
-        self.recv_queue = Queue()
         self.results = {}
 
-        self.ping_thread = None
-        self.ping_thread_event = Event()
+        self.ping_task = None
+        self.ping_task_event = asyncio.Event()
 
-        self.next_salt_thread = None
-        self.next_salt_thread_event = Event()
+        self.next_salt_task = None
+        self.next_salt_task_event = asyncio.Event()
 
-        self.net_worker_list = []
+        self.network_task = None
 
-        self.is_connected = Event()
+        self.is_connected = asyncio.Event()
 
-    def start(self):
+        self.loop = asyncio.get_event_loop()
+
+    async def start(self):
         while True:
             self.connection = Connection(
                 self.dc_id,
-                self.client.storage.test_mode(),
+                self.test_mode,
                 self.client.ipv6,
                 self.client.proxy
             )
 
             try:
-                self.connection.connect()
+                await self.connection.connect()
 
-                for i in range(self.NET_WORKERS):
-                    self.net_worker_list.append(
-                        Thread(
-                            target=self.net_worker,
-                            name="NetWorker#{}".format(i + 1)
-                        )
-                    )
+                self.network_task = self.loop.create_task(self.network_worker())
 
-                    self.net_worker_list[-1].start()
-
-                Thread(target=self.recv, name="RecvThread").start()
-
-                self.current_salt = FutureSalt(0, 0, self.INITIAL_SALT)
+                self.current_salt = FutureSalt(0, 0, Session.INITIAL_SALT)
                 self.current_salt = FutureSalt(
                     0, 0,
-                    self._send(
-                        functions.Ping(ping_id=0),
+                    (await self._send(
+                        raw.functions.Ping(ping_id=0),
                         timeout=self.START_TIMEOUT
-                    ).new_server_salt
+                    )).new_server_salt
                 )
-                self.current_salt = self._send(functions.GetFutureSalts(num=1), timeout=self.START_TIMEOUT).salts[0]
+                self.current_salt = (await self._send(
+                    raw.functions.GetFutureSalts(num=1),
+                    timeout=self.START_TIMEOUT)).salts[0]
 
-                self.next_salt_thread = Thread(target=self.next_salt, name="NextSaltThread")
-                self.next_salt_thread.start()
+                self.next_salt_task = self.loop.create_task(self.next_salt_worker())
 
                 if not self.is_cdn:
-                    self._send(
-                        functions.InvokeWithLayer(
+                    await self._send(
+                        raw.functions.InvokeWithLayer(
                             layer=layer,
-                            query=functions.InitConnection(
+                            query=raw.functions.InitConnection(
                                 api_id=self.client.api_id,
                                 app_version=self.client.app_version,
                                 device_model=self.client.device_model,
@@ -163,195 +157,161 @@ class Session:
                                 system_lang_code=self.client.lang_code,
                                 lang_code=self.client.lang_code,
                                 lang_pack="",
-                                query=functions.help.GetConfig(),
+                                query=raw.functions.help.GetConfig(),
                             )
                         ),
                         timeout=self.START_TIMEOUT
                     )
 
-                self.ping_thread = Thread(target=self.ping, name="PingThread")
-                self.ping_thread.start()
+                self.ping_task = self.loop.create_task(self.ping_worker())
 
-                log.info("Session initialized: Layer {}".format(layer))
-                log.info("Device: {} - {}".format(self.client.device_model, self.client.app_version))
-                log.info("System: {} ({})".format(self.client.system_version, self.client.lang_code.upper()))
+                log.info(f"Session initialized: Layer {layer}")
+                log.info(f"Device: {self.client.device_model} - {self.client.app_version}")
+                log.info(f"System: {self.client.system_version} ({self.client.lang_code.upper()})")
 
             except AuthKeyDuplicated as e:
-                self.stop()
+                await self.stop()
                 raise e
             except (OSError, TimeoutError, RPCError):
-                self.stop()
+                await self.stop()
             except Exception as e:
-                self.stop()
+                await self.stop()
                 raise e
             else:
                 break
 
         self.is_connected.set()
 
-        log.debug("Session started")
+        log.info("Session started")
 
-    def stop(self):
+    async def stop(self):
         self.is_connected.clear()
 
-        self.ping_thread_event.set()
-        self.next_salt_thread_event.set()
+        self.ping_task_event.set()
+        self.next_salt_task_event.set()
 
-        if self.ping_thread is not None:
-            self.ping_thread.join()
+        if self.ping_task is not None:
+            await self.ping_task
 
-        if self.next_salt_thread is not None:
-            self.next_salt_thread.join()
+        if self.next_salt_task is not None:
+            await self.next_salt_task
 
-        self.ping_thread_event.clear()
-        self.next_salt_thread_event.clear()
+        self.ping_task_event.clear()
+        self.next_salt_task_event.clear()
 
         self.connection.close()
 
-        for i in range(self.NET_WORKERS):
-            self.recv_queue.put(None)
-
-        for i in self.net_worker_list:
-            i.join()
-
-        self.net_worker_list.clear()
-        self.recv_queue.queue.clear()
+        if self.network_task:
+            await self.network_task
 
         for i in self.results.values():
             i.event.set()
 
         if not self.is_media and callable(self.client.disconnect_handler):
             try:
-                self.client.disconnect_handler(self.client)
+                await self.client.disconnect_handler(self.client)
             except Exception as e:
                 log.error(e, exc_info=True)
 
-        log.debug("Session stopped")
+        log.info("Session stopped")
 
-    def restart(self):
-        self.stop()
-        self.start()
+    async def restart(self):
+        await self.stop()
+        await self.start()
 
-    def pack(self, message: Message):
-        data = Long(self.current_salt.salt) + self.session_id + message.write()
-        padding = urandom(-(len(data) + 12) % 16 + 12)
+    async def handle_packet(self, packet):
+        if len(packet) <= self.EXECUTOR_SIZE_THRESHOLD:
+            data = mtproto.unpack(
+                BytesIO(packet),
+                self.session_id,
+                self.auth_key,
+                self.auth_key_id
+            )
+        else:
+            data = await self.loop.run_in_executor(
+                self.executor,
+                mtproto.unpack,
+                BytesIO(packet),
+                self.session_id,
+                self.auth_key,
+                self.auth_key_id
+            )
 
-        # 88 = 88 + 0 (outgoing message)
-        msg_key_large = sha256(self.auth_key[88: 88 + 32] + data + padding).digest()
-        msg_key = msg_key_large[8:24]
-        aes_key, aes_iv = KDF(self.auth_key, msg_key, True)
+        messages = (
+            data.body.messages
+            if isinstance(data.body, MsgContainer)
+            else [data]
+        )
 
-        return self.auth_key_id + msg_key + AES.ige256_encrypt(data + padding, aes_key, aes_iv)
+        log.debug(f"Received:\n{data}")
 
-    def unpack(self, b: BytesIO) -> Message:
-        assert b.read(8) == self.auth_key_id, b.getvalue()
+        for msg in messages:
+            if msg.seq_no == 0:
+                MsgId.set_server_time(msg.msg_id / (2 ** 32))
 
-        msg_key = b.read(16)
-        aes_key, aes_iv = KDF(self.auth_key, msg_key, False)
-        data = BytesIO(AES.ige256_decrypt(b.read(), aes_key, aes_iv))
-        data.read(8)
+            if msg.seq_no % 2 != 0:
+                if msg.msg_id in self.pending_acks:
+                    continue
+                else:
+                    self.pending_acks.add(msg.msg_id)
 
-        # https://core.telegram.org/mtproto/security_guidelines#checking-session-id
-        assert data.read(8) == self.session_id
+            if isinstance(msg.body, (raw.types.MsgDetailedInfo, raw.types.MsgNewDetailedInfo)):
+                self.pending_acks.add(msg.body.answer_msg_id)
+                continue
 
-        message = Message.read(data)
+            if isinstance(msg.body, raw.types.NewSessionCreated):
+                continue
 
-        # https://core.telegram.org/mtproto/security_guidelines#checking-sha256-hash-value-of-msg-key
-        # https://core.telegram.org/mtproto/security_guidelines#checking-message-length
-        # 96 = 88 + 8 (incoming message)
-        assert msg_key == sha256(self.auth_key[96:96 + 32] + data.getvalue()).digest()[8:24]
+            msg_id = None
 
-        # https://core.telegram.org/mtproto/security_guidelines#checking-msg-id
-        # TODO: check for lower msg_ids
-        assert message.msg_id % 2 != 0
+            if isinstance(msg.body, (raw.types.BadMsgNotification, raw.types.BadServerSalt)):
+                msg_id = msg.body.bad_msg_id
+            elif isinstance(msg.body, (FutureSalts, raw.types.RpcResult)):
+                msg_id = msg.body.req_msg_id
+            elif isinstance(msg.body, raw.types.Pong):
+                msg_id = msg.body.msg_id
+            else:
+                if self.client is not None:
+                    self.loop.create_task(self.client.handle_updates(msg.body))
 
-        return message
+            if msg_id in self.results:
+                self.results[msg_id].value = getattr(msg.body, "result", msg.body)
+                self.results[msg_id].event.set()
 
-    def net_worker(self):
-        name = threading.current_thread().name
-        log.debug("{} started".format(name))
+        if len(self.pending_acks) >= self.ACKS_THRESHOLD:
+            log.debug(f"Send {len(self.pending_acks)} acks")
+
+            try:
+                await self._send(raw.types.MsgsAck(msg_ids=list(self.pending_acks)), False)
+            except (OSError, TimeoutError):
+                pass
+            else:
+                self.pending_acks.clear()
+
+    async def ping_worker(self):
+        log.info("PingTask started")
 
         while True:
-            packet = self.recv_queue.get()
-
-            if packet is None:
+            try:
+                await asyncio.wait_for(self.ping_task_event.wait(), self.PING_INTERVAL)
+            except asyncio.TimeoutError:
+                pass
+            else:
                 break
 
             try:
-                data = self.unpack(BytesIO(packet))
-
-                messages = (
-                    data.body.messages
-                    if isinstance(data.body, MsgContainer)
-                    else [data]
+                await self._send(
+                    raw.functions.PingDelayDisconnect(
+                        ping_id=0, disconnect_delay=self.WAIT_TIMEOUT + 10
+                    ), False
                 )
-
-                log.debug("Received:\n{}".format(data))
-
-                for msg in messages:
-                    if msg.seq_no % 2 != 0:
-                        if msg.msg_id in self.pending_acks:
-                            continue
-                        else:
-                            self.pending_acks.add(msg.msg_id)
-
-                    if isinstance(msg.body, (types.MsgDetailedInfo, types.MsgNewDetailedInfo)):
-                        self.pending_acks.add(msg.body.answer_msg_id)
-                        continue
-
-                    if isinstance(msg.body, types.NewSessionCreated):
-                        continue
-
-                    msg_id = None
-
-                    if isinstance(msg.body, (types.BadMsgNotification, types.BadServerSalt)):
-                        msg_id = msg.body.bad_msg_id
-                    elif isinstance(msg.body, (core.FutureSalts, types.RpcResult)):
-                        msg_id = msg.body.req_msg_id
-                    elif isinstance(msg.body, types.Pong):
-                        msg_id = msg.body.msg_id
-                    else:
-                        if self.client is not None:
-                            self.client.updates_queue.put(msg.body)
-
-                    if msg_id in self.results:
-                        self.results[msg_id].value = getattr(msg.body, "result", msg.body)
-                        self.results[msg_id].event.set()
-
-                if len(self.pending_acks) >= self.ACKS_THRESHOLD:
-                    log.info("Send {} acks".format(len(self.pending_acks)))
-
-                    try:
-                        self._send(types.MsgsAck(msg_ids=list(self.pending_acks)), False)
-                    except (OSError, TimeoutError):
-                        pass
-                    else:
-                        self.pending_acks.clear()
-            except Exception as e:
-                log.error(e, exc_info=True)
-
-        log.debug("{} stopped".format(name))
-
-    def ping(self):
-        log.debug("PingThread started")
-
-        while True:
-            self.ping_thread_event.wait(self.PING_INTERVAL)
-
-            if self.ping_thread_event.is_set():
-                break
-
-            try:
-                self._send(functions.PingDelayDisconnect(
-                    ping_id=0, disconnect_delay=self.WAIT_TIMEOUT + 10
-                ), False)
             except (OSError, TimeoutError, RPCError):
                 pass
 
-        log.debug("PingThread stopped")
+        log.info("PingTask stopped")
 
-    def next_salt(self):
-        log.debug("NextSaltThread started")
+    async def next_salt_worker(self):
+        log.info("NextSaltTask started")
 
         while True:
             now = datetime.now()
@@ -361,90 +321,115 @@ class Session:
             valid_until = datetime.fromtimestamp(self.current_salt.valid_until)
             dt = (valid_until - now).total_seconds() - 900
 
-            log.debug("Current salt: {} | Next salt in {:.0f}m {:.0f}s ({})".format(
-                self.current_salt.salt,
-                dt // 60,
-                dt % 60,
+            log.info("Next salt in {:.0f}m {:.0f}s ({})".format(
+                dt // 60, dt % 60,
                 now + timedelta(seconds=dt)
             ))
 
-            self.next_salt_thread_event.wait(dt)
-
-            if self.next_salt_thread_event.is_set():
+            try:
+                await asyncio.wait_for(self.next_salt_task_event.wait(), dt)
+            except asyncio.TimeoutError:
+                pass
+            else:
                 break
 
             try:
-                self.current_salt = self._send(functions.GetFutureSalts(num=1)).salts[0]
+                self.current_salt = (await self._send(raw.functions.GetFutureSalts(num=1))).salts[0]
             except (OSError, TimeoutError, RPCError):
                 self.connection.close()
                 break
 
-        log.debug("NextSaltThread stopped")
+        log.info("NextSaltTask stopped")
 
-    def recv(self):
-        log.debug("RecvThread started")
+    async def network_worker(self):
+        log.info("NetworkTask started")
 
         while True:
-            packet = self.connection.recv()
+            packet = await self.connection.recv()
 
             if packet is None or len(packet) == 4:
                 if packet:
-                    log.warning("Server sent \"{}\"".format(Int.read(BytesIO(packet))))
+                    log.warning(f'Server sent "{Int.read(BytesIO(packet))}"')
 
                 if self.is_connected.is_set():
-                    Thread(target=self.restart, name="RestartThread").start()
+                    self.loop.create_task(self.restart())
+
                 break
 
-            self.recv_queue.put(packet)
+            self.loop.create_task(self.handle_packet(packet))
 
-        log.debug("RecvThread stopped")
+        log.info("NetworkTask stopped")
 
-    def _send(self, data: TLObject, wait_response: bool = True, timeout: float = WAIT_TIMEOUT):
+    async def _send(self, data: TLObject, wait_response: bool = True, timeout: float = WAIT_TIMEOUT):
         message = self.msg_factory(data)
         msg_id = message.msg_id
 
         if wait_response:
             self.results[msg_id] = Result()
 
-        log.debug("Sent:\n{}".format(message))
+        log.debug(f"Sent:\n{message}")
 
-        payload = self.pack(message)
+        if len(message) <= self.EXECUTOR_SIZE_THRESHOLD:
+            payload = mtproto.pack(
+                message,
+                self.current_salt.salt,
+                self.session_id,
+                self.auth_key,
+                self.auth_key_id
+            )
+        else:
+            payload = await self.loop.run_in_executor(
+                self.executor,
+                mtproto.pack,
+                message,
+                self.current_salt.salt,
+                self.session_id,
+                self.auth_key,
+                self.auth_key_id
+            )
 
         try:
-            self.connection.send(payload)
+            await self.connection.send(payload)
         except OSError as e:
             self.results.pop(msg_id, None)
             raise e
 
         if wait_response:
-            self.results[msg_id].event.wait(timeout)
-            result = self.results.pop(msg_id).value
+            try:
+                await asyncio.wait_for(self.results[msg_id].event.wait(), timeout)
+            except asyncio.TimeoutError:
+                pass
+            finally:
+                result = self.results.pop(msg_id).value
 
             if result is None:
                 raise TimeoutError
-            elif isinstance(result, types.RpcError):
-                if isinstance(data, (functions.InvokeWithoutUpdates, functions.InvokeWithTakeout)):
+            elif isinstance(result, raw.types.RpcError):
+                if isinstance(data, (raw.functions.InvokeWithoutUpdates, raw.functions.InvokeWithTakeout)):
                     data = data.query
 
                 RPCError.raise_it(result, type(data))
-            elif isinstance(result, types.BadMsgNotification):
+            elif isinstance(result, raw.types.BadMsgNotification):
                 raise Exception(self.BAD_MSG_DESCRIPTION.get(
                     result.error_code,
-                    "Error code {}".format(result.error_code)
+                    f"Error code {result.error_code}"
                 ))
             else:
                 return result
 
-    def send(
+    async def send(
         self,
         data: TLObject,
         retries: int = MAX_RETRIES,
         timeout: float = WAIT_TIMEOUT,
         sleep_threshold: float = SLEEP_THRESHOLD
     ):
-        self.is_connected.wait(self.WAIT_TIMEOUT)
+        try:
+            await asyncio.wait_for(self.is_connected.wait(), self.WAIT_TIMEOUT)
+        except asyncio.TimeoutError:
+            pass
 
-        if isinstance(data, (functions.InvokeWithoutUpdates, functions.InvokeWithTakeout)):
+        if isinstance(data, (raw.functions.InvokeWithoutUpdates, raw.functions.InvokeWithTakeout)):
             query = data.query
         else:
             query = data
@@ -453,26 +438,23 @@ class Session:
 
         while True:
             try:
-                return self._send(data, timeout=timeout)
+                return await self._send(data, timeout=timeout)
             except FloodWait as e:
                 amount = e.x
 
-                if amount > sleep_threshold:
+                if amount > sleep_threshold >= 0:
                     raise
 
-                log.warning('[{}] Sleeping for {}s (required by "{}")'.format(
-                    self.client.session_name, amount, query))
+                log.warning(f'[{self.client.session_name}] Sleeping for {amount}s (required by "{query}")')
 
-                time.sleep(amount)
+                await asyncio.sleep(amount)
             except (OSError, TimeoutError, InternalServerError) as e:
                 if retries == 0:
                     raise e from None
 
                 (log.warning if retries < 2 else log.info)(
-                    '[{}] Retrying "{}" due to {}'.format(
-                        Session.MAX_RETRIES - retries + 1,
-                        query, e))
+                    f'[{Session.MAX_RETRIES - retries + 1}] Retrying "{query}" due to {e}')
 
-                time.sleep(0.5)
+                await asyncio.sleep(0.5)
 
-                return self.send(data, retries - 1, timeout)
+                return await self.send(data, retries - 1, timeout)
