@@ -19,8 +19,8 @@
 import logging
 import os
 
+from pyrogram.crypto import aes
 from .tcp import TCP
-from ....crypto.aes import AES
 
 log = logging.getLogger(__name__)
 
@@ -34,8 +34,8 @@ class TCPAbridgedO(TCP):
         self.encrypt = None
         self.decrypt = None
 
-    def connect(self, address: tuple):
-        super().connect(address)
+    async def connect(self, address: tuple):
+        await super().connect(address)
 
         while True:
             nonce = bytearray(os.urandom(64))
@@ -49,15 +49,15 @@ class TCPAbridgedO(TCP):
         self.encrypt = (nonce[8:40], nonce[40:56], bytearray(1))
         self.decrypt = (temp[0:32], temp[32:48], bytearray(1))
 
-        nonce[56:64] = AES.ctr256_encrypt(nonce, *self.encrypt)[56:64]
+        nonce[56:64] = aes.ctr256_encrypt(nonce, *self.encrypt)[56:64]
 
-        super().sendall(nonce)
+        await super().send(nonce)
 
-    def sendall(self, data: bytes, *args):
+    async def send(self, data: bytes, *args):
         length = len(data) // 4
 
-        super().sendall(
-            AES.ctr256_encrypt(
+        await super().send(
+            aes.ctr256_encrypt(
                 (bytes([length])
                  if length <= 126
                  else b"\x7f" + length.to_bytes(3, "little"))
@@ -66,25 +66,25 @@ class TCPAbridgedO(TCP):
             )
         )
 
-    def recvall(self, length: int = 0) -> bytes or None:
-        length = super().recvall(1)
+    async def recv(self, length: int = 0) -> bytes or None:
+        length = await super().recv(1)
 
         if length is None:
             return None
 
-        length = AES.ctr256_decrypt(length, *self.decrypt)
+        length = aes.ctr256_decrypt(length, *self.decrypt)
 
         if length == b"\x7f":
-            length = super().recvall(3)
+            length = await super().recv(3)
 
             if length is None:
                 return None
 
-            length = AES.ctr256_decrypt(length, *self.decrypt)
+            length = aes.ctr256_decrypt(length, *self.decrypt)
 
-        data = super().recvall(int.from_bytes(length, "little") * 4)
+        data = await super().recv(int.from_bytes(length, "little") * 4)
 
         if data is None:
             return None
 
-        return AES.ctr256_decrypt(data, *self.decrypt)
+        return aes.ctr256_decrypt(data, *self.decrypt)
